styled tasks in calendar
