+++ src/components/Calendar/ChoosedMonth/ChoosedMonth.styled.jsx
@@ -4,6 +4,7 @@ import styled from '@emotion/styled';
 export const Table = styled.div`
     display: grid;
     grid-template-columns: repeat(7, 1fr);
+    border-radius: 8px;
     ${props => props.fullheight}
     ${props =>
         props.fullheight &&
@@ -13,15 +14,17 @@ export const Table = styled.div`
         justify-content: space-between;
         flex-direction: column;
         align-items: flex-end;
-        padding: 8px;
+        padding: 4px;
         overflow: hidden;
         // max-width: 165px;
         height: 95px;
         background-color: white;
         border: 1px solid rgba(220, 227, 229, 0.8);
+        
             @media screen and (min-width: ${({ theme }) => theme.breakpoints.m}) {
               max-width: 155px;
-              height: 125px;
+              height: 134px;
+              padding: 8px;
             };
         span {
             font-family: Inter;
